server/requirements.txt
openenv-core>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0
numpy>=1.24.0
